+++ CHANGES.rst
@@ -14,6 +14,8 @@ Features added
 * The embedded C code comments that show the original source code
   can be discarded with the new directive ``emit_code_comments=False``.
 
+* Cpdef enums are now first-class iterable, callable types in Python.
+
 * Posix declarations for DLL loading and stdio extensions were added.
   Patch by Lars Buitinck.